requirements.txt
streamlit
pandas
sqlalchemy
python-dotenv
langchain-openai
matplotlib